+++ packages/login_client/lib/src/credentials_storage/in_memory_credentials_storage.dart
@@ -18,10 +18,10 @@ import 'credentials_storage.dart';
 
 /// A [CredentialsStorage] implementation storing credentials in the memory.
 class InMemoryCredentialsStorage extends CredentialsStorage {
-  oauth2.Credentials _credentials;
+  oauth2.Credentials? _credentials;
 
   @override
-  Future<oauth2.Credentials> read() {
+  Future<oauth2.Credentials?> read() {
     return Future.value(_credentials);
   }