+++ packages/login_client/lib/src/oauth_settings.dart
@@ -12,18 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'package:meta/meta.dart';
-
 /// OAuth2 authorization requests settings.
 class OAuthSettings {
   /// Creates the [OAuthSettings].
   const OAuthSettings({
-    @required this.authorizationUri,
-    @required this.clientId,
+    required this.authorizationUri,
+    required this.clientId,
     this.clientSecret = '',
     this.scopes = const [],
-  })  : assert(clientId != null),
-        assert(clientSecret != null);
+  });
 
   /// The [`authorization endpoint`](https://tools.ietf.org/html/rfc6749#section-3.1)
   /// from the RFC 6749.