+++ packages/login_client/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0
+
+- **Breaking:** Migrate to null-safety.
+- **Breaking:** Bump minimum Dart version to 2.12.
+- **Breaking**: Remove deprecated `credentialsChangedCallback` parameter from `LoginClient` constructor.
+
 # 1.1.1
 
 - Fix closing default HTTP client after logging out.